pyproject.toml
[tool.poetry]
name = "catholic-drive-server"
version = "0.1.0"
description = ""
authors = ["Moonsik Park <moonsik.park@estsoft.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.1"
django-rest-framework = "^0.1.0"
pgvector = "^0.3.2"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
